dashboard/requirements.txt
shiny
shinyswatch
shinywidgets
geopandas
pyproj
pandas
rioxarray
xarray
numpy
pyarrow
matplotlib 
python-dotenv
ipyleaflet
fsspec
gcsfs
google-cloud-storage